[refactor] 구조 수정 및 Optimize
