+++ src/main/java/com/example/realq/domain/bookmark/entity/BookmarkRegion.java
@@ -1,6 +1,6 @@
 package com.example.realq.domain.bookmark.entity;
 
-import com.example.realq.domain.region.entity.Region;
+import com.example.realq.domain.realtime.region.entity.Region;
 import com.example.realq.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
