+++ src/main/java/com/example/realq/domain/bookmark/repository/BookmarkRegionRepository.java
@@ -1,7 +1,6 @@
 package com.example.realq.domain.bookmark.repository;
 
 import com.example.realq.domain.bookmark.entity.BookmarkRegion;
-import com.example.realq.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
