+++ src/main/java/com/example/realq/domain/realtime/region/dto/response/RegionBody.java
@@ -1,0 +1,7 @@
+package com.example.realq.domain.realtime.region.dto.response;
+
+import java.util.List;
+
+public record RegionBody(
+        List<RegionItem> items
+) {}
