+++ src/main/java/com/example/realq/domain/realtime/region/dto/response/RegionResponseBody.java
@@ -1,0 +1,5 @@
+package com.example.realq.domain.realtime.region.dto.response;
+
+public record RegionResponseBody(
+        RegionBody body
+) {}
