+++ src/main/java/com/example/realq/domain/realtime/region/entity/Region.java
@@ -1,0 +1,16 @@
+package com.example.realq.domain.realtime.region.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+
+@Entity
+@Getter
+@Table(name = "region")
+public class Region {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+}
