+++ src/main/java/com/example/realq/domain/realtime/region/repository/RegionRepository.java
@@ -1,0 +1,8 @@
+package com.example.realq.domain.realtime.region.repository;
+
+import com.example.realq.domain.realtime.region.entity.Region;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RegionRepository extends JpaRepository<Region, Long> {
+
+}
