+++ src/main/java/com/example/realq/domain/realtime/station/dto/response/StationResponseBody.java
@@ -1,0 +1,9 @@
+package com.example.realq.domain.realtime.station.dto.response;
+
+public record StationResponseBody(
+        StationBody body
+) {
+    public record StationBody(
+            StationItem[] items
+    ) {}
+}
