+++ src/main/java/com/example/realq/domain/realtime/station/dto/response/StationWrapper.java
@@ -1,0 +1,5 @@
+package com.example.realq.domain.realtime.station.dto.response;
+
+public record StationWrapper(
+        StationResponseBody response
+) {}
