+++ src/main/java/com/example/realq/domain/realtime/station/entity/Station.java
@@ -1,0 +1,23 @@
+package com.example.realq.domain.realtime.station.entity;
+
+import com.example.realq.domain.realtime.region.entity.Region;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@Table(name = "station")
+public class Station {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    @ManyToOne
+    @JoinColumn(name = "region_id")
+    private Region region;
+}
