--- src/main/java/com/example/realq/domain/region/dto/response/RegionBody.java
@@ -1,7 +1,0 @@
-package com.example.realq.domain.region.dto.response;
-
-import java.util.List;
-
-public record RegionBody(
-        List<RegionItem> items
-) {}
