--- src/main/java/com/example/realq/domain/region/dto/response/RegionWrapper.java
@@ -1,5 +1,0 @@
-package com.example.realq.domain.region.dto.response;
-
-public record RegionWrapper(
-        RegionResponseBody response
-) {}
