--- src/main/java/com/example/realq/domain/region/entity/Region.java
@@ -1,16 +1,0 @@
-package com.example.realq.domain.region.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-
-@Entity
-@Getter
-@Table(name = "region")
-public class Region {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-}
