--- src/main/java/com/example/realq/domain/region/repository/RegionRepository.java
@@ -1,8 +1,0 @@
-package com.example.realq.domain.region.repository;
-
-import com.example.realq.domain.region.entity.Region;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RegionRepository extends JpaRepository<Region, Long> {
-
-}
