--- src/main/java/com/example/realq/domain/station/dto/response/StationResponse.java
@@ -1,12 +1,0 @@
-package com.example.realq.domain.station.dto.response;
-
-public record StationResponse(
-        String stationName,
-        String dataTime,
-        double pm10,
-        double pm25,
-        double o3,
-        double no2,
-        double co,
-        double so2
-) {}
