--- src/main/java/com/example/realq/domain/station/dto/response/StationResponseBody.java
@@ -1,9 +1,0 @@
-package com.example.realq.domain.station.dto.response;
-
-public record StationResponseBody(
-        StationBody body
-) {
-    public record StationBody(
-            StationItem[] items
-    ) {}
-}
