--- src/main/java/com/example/realq/domain/station/dto/response/StationWrapper.java
@@ -1,5 +1,0 @@
-package com.example.realq.domain.station.dto.response;
-
-public record StationWrapper(
-        StationResponseBody response
-) {}
