--- src/main/java/com/example/realq/domain/station/repository/StationRepository.java
@@ -1,7 +1,0 @@
-package com.example.realq.domain.station.repository;
-
-import com.example.realq.domain.station.entity.Station;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StationRepository extends JpaRepository<Station, Long> {
-}
